refactor: 카드 프리뷰 관련 파일 분리
